add sample keyword to sample package name to avoid namespace conflict… (#155)
* add sample keyword to sample package name to avoid namespace conflict with starter packages

* reformat
