--- documentdb/spring-data-azure-documentdb-sample/src/main/java/com/microsoft/azure/sample/AppConfiguration.java
@@ -1,37 +1,0 @@
-/**
- * Copyright (c) Microsoft Corporation. All rights reserved.
- * Licensed under the MIT License. See LICENSE in the project root for
- * license information.
- */
-
-package com.microsoft.azure.sample;
-
-import com.microsoft.azure.documentdb.ConnectionPolicy;
-import com.microsoft.azure.documentdb.ConsistencyLevel;
-import com.microsoft.azure.documentdb.DocumentClient;
-import com.microsoft.azure.spring.data.documentdb.config.AbstractDocumentDbConfiguration;
-import com.microsoft.azure.spring.data.documentdb.repository.config.EnableDocumentDbRepositories;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-@EnableDocumentDbRepositories
-public class AppConfiguration extends AbstractDocumentDbConfiguration {
-
-    @Value("${azure.documentdb.uri}")
-    private String uri;
-
-    @Value("${azure.documentdb.key}")
-    private String key;
-
-    @Value("${azure.documentdb.database}")
-    private String dbName;
-
-    public DocumentClient documentClient() {
-        return new DocumentClient(uri, key, ConnectionPolicy.GetDefault(), ConsistencyLevel.Session);
-    }
-
-    public String getDatabase() {
-        return dbName;
-    }
-}
