--- documentdb/spring-data-azure-documentdb-sample/src/main/java/com/microsoft/azure/sample/User.java
@@ -1,56 +1,0 @@
-/**
- * Copyright (c) Microsoft Corporation. All rights reserved.
- * Licensed under the MIT License. See LICENSE in the project root for
- * license information.
- */
-
-package com.microsoft.azure.sample;
-
-import com.microsoft.azure.spring.data.documentdb.core.mapping.Document;
-import org.springframework.data.annotation.Id;
-
-@Document(collection = "mycollection")
-public class User {
-    @Id
-    private String emailAddress;
-    private String firstName;
-    private String lastName;
-
-    public User(String emailAddress, String firstName, String lastName) {
-        this.emailAddress = emailAddress;
-        this.firstName = firstName;
-        this.lastName = lastName;
-    }
-
-    public String getEmailAddress() {
-        return emailAddress;
-    }
-
-    public void setEmailAddress(String emailAddress) {
-        this.emailAddress = emailAddress;
-    }
-
-
-    public String getFirstName() {
-        return firstName;
-    }
-
-    public void setFirstName(String firstName) {
-        this.firstName = firstName;
-    }
-
-    public String getLastName() {
-        return lastName;
-    }
-
-    public void setLastName(String lastName) {
-        this.lastName = lastName;
-    }
-
-
-    @Override
-    public String toString() {
-        return String.format("%s: %s %s", emailAddress, firstName, lastName);
-    }
-}
-
