+++ documentdb/spring-data-azure-documentdb-sample/src/main/java/com/microsoft/azure/spring/data/documentdb/sample/SampleApplication.java
@@ -1,0 +1,44 @@
+/**
+ * Copyright (c) Microsoft Corporation. All rights reserved.
+ * Licensed under the MIT License. See LICENSE in the project root for
+ * license information.
+ */
+
+package com.microsoft.azure.spring.data.documentdb.sample;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.CommandLineRunner;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.util.Assert;
+
+@SpringBootApplication
+public class SampleApplication implements CommandLineRunner {
+
+    private static final Logger LOGGER = LoggerFactory.getLogger(SampleApplication.class);
+
+    @Autowired
+    private UserRepository repository;
+
+    public static void main(String[] args) {
+        SpringApplication.run(SampleApplication.class, args);
+    }
+
+    public void run(String... var1) throws Exception {
+
+        final User testUser = new User("test@test.com", "testFirstName", "testLastName");
+
+        repository.deleteAll();
+        repository.save(testUser);
+
+        final User result = repository.findOne(testUser.getEmailAddress());
+
+        Assert.state(result.getFirstName().equals(testUser.getFirstName()), "query result firstName doesn't match!");
+        Assert.state(result.getLastName().equals(testUser.getLastName()), "query result lastName doesn't match!");
+
+        LOGGER.info("findOne in User collection get result: {}", result.toString());
+        System.out.println("findOne in User collection get result:" + result.toString());
+    }
+}
