+++ documentdb/spring-data-azure-documentdb-sample/src/main/java/com/microsoft/azure/spring/data/documentdb/sample/UserRepository.java
@@ -1,0 +1,15 @@
+/**
+ * Copyright (c) Microsoft Corporation. All rights reserved.
+ * Licensed under the MIT License. See LICENSE in the project root for
+ * license information.
+ */
+
+package com.microsoft.azure.spring.data.documentdb.sample;
+
+import com.microsoft.azure.spring.data.documentdb.repository.DocumentDbRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserRepository extends DocumentDbRepository<User, String> {
+}
+
