--- microsoft-graph/microsoft-graph-spring-social-starter-sample/src/main/java/sample/microsoft/graph/custom/Contact.java
@@ -1,36 +1,0 @@
-/**
- * Copyright (c) Microsoft Corporation. All rights reserved.
- * Licensed under the MIT License. See LICENSE in the project root for
- * license information.
- */
-
-package sample.microsoft.graph.custom;
-
-// Please be noted this is not a complete representation of the JSON object returned by Microsoft Graph.
-public class Contact {
-
-    private String displayName;
-    private String mobilePhone;
-
-    /**
-     * The Display Name.
-     */
-    public String getDisplayName() {
-        return displayName;
-    }
-
-    public void setDisplayName(String displayName) {
-        this.displayName = displayName;
-    }
-
-    /**
-     * The Mobile Phone.
-     */
-    public String getMobilePhone() {
-        return mobilePhone;
-    }
-
-    public void setMobilePhone(String mobilePhone) {
-        this.mobilePhone = mobilePhone;
-    }
-}
